Reviewed mpf_codec_t and mpf_stream_t entities and relationship between them.
Codec as a manipulator (encode,decode,dissect) is not a property of stream. Stream holds capabilities and active codec descriptor(s). Actual codecs (PCMU, PCMA, L16) are registered at startup and passed to encoder and decoder objects when/where needed.
Linear PCM (LPCM, but not L16) is just an internal descriptor, it's not a codec.

Everything should work as is basis. The only user level interface change is in stream_open() method, which is optional and probably is not actually used. See demos for more info.

git-svn-id: http://unimrcp.googlecode.com/svn@1133 f001bc3a-424a-0410-80a0-a715b8f413a8