Allow a pool/factory of signaling agents to be defined with the MRCPv1 and/or MRCPv2 profiles. Signaling agents are specified using a comma-separated list as follows:
<sip-uac>SIP-Agent-1,SIP-Agent-2,SIP-Agent-3</sip-uac>

and are selected for a new session using a simple round-robin scheduling.

The following definition is still legitimate and is equivalent to the pool/factory of one signaling agent.
<sip-uac>SIP-Agent-1</sip-uac>

This technique allows better utilization of resources under a heavy load with hundreds of simultaneous sessions. The primary target (bottleneck) is Sofia-SIP su_root task.

Nothing has to be changed.

git-svn-id: http://unimrcp.googlecode.com/svn@2060 f001bc3a-424a-0410-80a0-a715b8f413a8