Separated var and data directories.
The datadir contains read-only data installed with and used by applications. While the vardir is supposed to hold files being created and modified on run-time.

On Windows, the "prepare" project needs to be built in order to create an empty var directory (or just create it manually next to the other subdirs).

git-svn-id: http://unimrcp.googlecode.com/svn@2193 f001bc3a-424a-0410-80a0-a715b8f413a8